initial impl. (translation)
