+++ src/app/pages/create-game/create-game-list/create-game-list.page.html
@@ -128,7 +128,7 @@
         >{{ "CreateTasks.chooseComponents" | translate }}
         <ion-icon
           name="information-circle"
-          (click)="showPopover($event, 'Popover.chooseComponent')"
+          (click)="utilService.showPopover_i18n($event, 'Popover.chooseComponent')"
         >
         </ion-icon>
       </ion-text>
