+++ src/app/pages/create-game/create-game-virtual-list/create-game-virtual-list.page.html
@@ -62,7 +62,7 @@
   <ion-card color="light">
     <ion-card-content>
       <ion-text>{{ "CreateTasks.chooseComponents" | translate }}
-        <ion-icon name="information-circle" (click)="showPopover($event, 'Popover.chooseComponent')">
+        <ion-icon name="information-circle" (click)="utilService.showPopover_i18n($event, 'Popover.chooseComponent')">
         </ion-icon>
       </ion-text>
       <ion-item color="light" lines="none" *ngIf="game">
