+++ src/assets/i18n/de.json
@@ -92,7 +92,13 @@
     "logInToAccess": "Sie müssen sich anmelden, um Zugriff zu haben",
     "players": "Spieler",
     "tasks": "Aufgaben",
-    "task": "Aufgabe"
+    "task": "Aufgabe",
+    "gameEnv":"Spielumgebung",
+    "gameEnv_info":"Wähle ob Sie GeoGami draußen (Realraum) oder drinnen (virtuelle Umgebung) spielen möchten.",
+    "gameMode":"Spielemodus",
+    "gameMode_info":"Wählen ob Sie GeoGami einzeln oder in einer Gruppe spielen möchten. Die Mindestanzahl an Spielern im Mehrspielermodus beträgt 2.",
+    "singlePlayer":"Einzelspieler",
+    "multiplayer":"Mehrspieler"
   },
   "CreateTasks": {
     "combineTasks": "Liste der erstellten Aufgaben:",
@@ -116,7 +122,7 @@
   },
   "Start": {
     "playMode": "Spielen",
-    "playModeDescription": "Wähle ein GeoGami aus der Spieleliste.",
+    "playModeDescription": "Wähle ein Spiel aus der Spieleliste.",
     "developMode": "Entwickeln",
     "developModeDescription": "Erstelle oder bearbeite ein GeoGami.",
     "EvaluateMode": "Auswerten",
@@ -188,7 +194,8 @@
     "photoDirection": "Foto der zu suchenden Richtung",
     "photoTask": "Foto zur Aufgabe",
     "correctAnswer": "Die korrekte Anwort:",
-    "selectGameType":"Spieltyp auswählen"
+    "selectGameType":"Spieltyp auswählen",
+    "setAvatarPosition":"Anfangsposition und Richtung des Avatars festlegen"
   },
   "EditGame": {
     "titleEditGame": "Bearbeite dein GeoGami",
@@ -229,7 +236,9 @@
     "gameNotFound":"Das Spiel konnte nicht gefunden werden.",
     "deleteGame": "Spiel löschen",
     "deleteGameMsg": "Bist du sicher, dass du dieses Spiel löschen möchtest?",
-    "useExternalVEApp":"Externe Virtual Environment App verwenden (WinOS/MacOS)"
+    "useExternalVEApp":"Externe Virtual Environment App verwenden (WinOS/MacOS)",
+    "enterName":"Gib deinen Namen ein",
+    "askPlayerToScanQRCode":"Bitte die Spieler, den untenstehenden QR-Code zu scannen, um am Spiel teilzunehmen."
   },
   "Showroom": {
     "titleShowroom": "Lerne GeoGami kennen",
